refactor: move methods related to forming ApiGatewaySpec to the class
